ssf/requirements.txt
argparse==1.4.0
packaging==23.1
pyyaml==5.4.1
fastapi==0.95.2
uvicorn==0.22.0
prometheus_fastapi_instrumentator==6.0.0
prometheus-client==0.9.0
python-multipart==0.0.6
typing-extensions==4.6.2
requests==2.31.0
wheel>=0.38.1
